refactor(ui): rename components, move components to respective directories
Moving the UI components to their respective directories for better organization and to reduce nesting.

BREAKING CHANGE:
- Many component names have been changed due to the refactoring.
