+++ src/components/home/NpcSection.jsx
