+++ src/components/npc/ListItem.jsx
