+++ src/components/npc/QuestBadge.jsx
