+++ src/pages/Home.jsx
@@ -1,4 +1,4 @@
-import HomeContent from "../organisme/HomeContent";
+import HomeContent from "../components/home/HomeContent";
 
 export default function Home() {
   return (
