+++ src/pages/NpcDetail.jsx
@@ -1,0 +1,9 @@
+import NpcContent from "../components/npc/NpcContent";
+
+export default function NpcDetail() {
+  return (
+    <div className="container mx-auto min-h-screen p-4">
+      <NpcContent />
+    </div>
+  );
+}
